// mc-utils/src/error/errors.interface.ts
/**
 * Copyright {Project Name} 2022 All Rights Reserved.
 * Author: Miguel Correlo
 * Node module: @lib/utils
 * This file is licensed under the MIT License.
 * License text available at https://opensource.org/licenses/MIT
 */

export enum HttpStatusCode {
  OK = 200,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  NOT_FOUND = 404,
  INTERNAL_SERVER = 500,
}